pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shadowcorr"
version = "0.1.0"
description = "ShadowCorr: Correspondence via Volumetric Consensus for Multi-View 3D Segments"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "tensorboard>=2.14.0",
  "numpy",
  "scipy",
  "scikit-learn",
  "tqdm",
  "pandas",
  "open3d",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["shadowcorr*"]

[tool.setuptools.package-data]
shadowcorr = ["conf/*.yaml"]

[project.scripts]
shadowcorr            = "shadowcorr.cli.main:main"
shadowcorr-train      = "shadowcorr.train_app:main"
shadowcorr-eval       = "shadowcorr.eval_app:main"
shadowcorr-preprocess = "shadowcorr.preprocess_app:main"
